auth0 node version, psql checks, cleanup
Lock the node version to match common version we are developing in (v10
vs v14:latest), locking major and minor version.  Up to the team how
minor they want to lock.

Psql check logic will help fix issues where there is not ready database
when the auth0 service is up and ready to do migrations. We can check
for a ready DB before starting node JS scripts that trigger migrations.

This is not for prod, so it may want to use other logic otherwise.  For
testing and dev, it's an easy fix.