add silid e2e d-compose & all containers
Previous `docker-compose.e2e.yml` was in the silid-server context only,
and required manual starting of the client build server.  DB migrations
would sometimes not work.  Created this compose to start all 4x required
service containers to run cypress tests.  Also added PM2 to the client
service since it sends a premature exit status as the JS scripts are
running.  PM2 watches the app, and is only needed for testing.

This compose was built for Linux using `network_mode: host`, and is not
supported in macOS.  It would require changing the behavior of how the
individual JS service apps talk to each other and hand-off communication
to use native Docker port mapping for execution on macOS.  This requires
an indepth knowledge of the application stack.